ci/emu_mps/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emu-mps"
description = "Pasqal MPS based pulse emulator built on PyTorch"
readme = "../../README.md"
authors = [
    { name = "Anton Quelle", email = "anton.quelle@pasqal.com"},
    { name = "Mauro Mendizabal", email="mauro.mendizabal-pico@pasqal.com" },
    { name = "Stefano Grava", email="stefano.grava@pasqal.com" },
    { name = "Pablo Le Henaff", email="pablo.le-henaff@pasqal.com" }
]
url="https://gitlab.pasqal.com/emulation/rydberg-atoms/emu-ct"
requires-python = ">=3.10"
license = {text = "Proprietary"}
classifiers=[
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
  "pulser-core==0.20.*",
  "torch>=2.4.0",
  "emu-base==1.1.1"]

[tool.hatch.version]
path = "../../emu_mps/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true
allow-ambiguous-features = true

[tool.hatch.envs.docs]
dependencies = [
]

[tool.hatch.build.targets.sdist]
packages = ["emu_mps"]

[tool.hatch.build.targets.sdist.force-include]
"../../emu_mps" = "emu_mps"

[tool.hatch.build.targets.wheel]
packages = ["emu_mps"]

[tool.hatch.build.targets.wheel.force-include]
"../../emu_mps" = "emu_mps"

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black] #sometimes black uses this pyproject, rather than the other one for config.
line-length = 90
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
